refactor: move std/streams.ts to lib/
