+++ lib/streams_test.ts
